Refactor frame handles and tighten evidence hashes
Extracted large inline logic into reusable modules: ScatterplotMatrix cell components were moved to `scatterplotMatrixCells.tsx`, evidence envelope input/validation logic was split into dedicated helpers, and Stream XY/Ordinal/Geo imperative ref APIs were moved into `use*FrameHandle` hooks. Added a shared `renderedSceneHash` utility and updated both server rendering and MCP rendering to compute/bind scene hashes from the final SVG output, improving evidence consistency after post-render transformations. Also removed the ScatterplotMatrix file-size grandfather entry now that the file has been split.
