+++ src/components/evidence/chartEvidenceEnvelope.ts
@@ -35,6 +35,16 @@ import {
 } from "../artifact/types"
 import type { RenderEvidence } from "../server/renderEvidence"
 import { stableEvidenceHash } from "./stableJsonHash"
+import {
+  normalizeSourceRecords,
+  redactNavigationTree,
+  redactProfileForEnvelope
+} from "./evidenceEnvelopeInput"
+import {
+  validateEnvelopeAccessibilityAudit,
+  validateEnvelopeArtifactAttachment,
+  validateEnvelopeSceneHash
+} from "./evidenceEnvelopeValidation"
 export { stableEvidenceHash } from "./stableJsonHash"
 
 export const CHART_EVIDENCE_ENVELOPE_VERSION = 1 as const
@@ -265,111 +275,6 @@ function inferRenderParity(input: {
   return input.observed === input.intended ? "match" : "unknown"
 }
 
-/**
- * Remove raw row/network/hierarchy/geo payloads from a data profile before
- * placing it in a portable envelope. Shape and aggregate evidence remain;
- * caller-owned source records do not.
- */
-const PRIVATE_PROFILE_KEYS = new Set([
-  "data",
-  "rawInput",
-  "sample",
-  "numericFields",
-  "network",
-  "hierarchy",
-  "geo",
-  "topValues",
-  "distinctValues"
-])
-
-function redactProfileForEnvelope(profile: unknown): unknown {
-  if (Array.isArray(profile)) return profile.map(redactProfileForEnvelope)
-  if (!profile || typeof profile !== "object") return profile
-  const output: Record<string, unknown> = {}
-  for (const [key, value] of Object.entries(
-    profile as Record<string, unknown>
-  )) {
-    if (value === undefined || PRIVATE_PROFILE_KEYS.has(key)) continue
-    output[key] = redactProfileForEnvelope(value)
-  }
-  return output
-}
-
-/** Recursively remove raw source records from a navigation tree. */
-function redactNavigationTree(tree: unknown): unknown {
-  if (Array.isArray(tree)) return tree.map(redactNavigationTree)
-  if (!tree || typeof tree !== "object") return tree
-  const output = { ...(tree as Record<string, unknown>) }
-  delete output.datum
-  if (Array.isArray(output.children)) {
-    output.children = output.children.map(redactNavigationTree)
-  }
-  return output
-}
-
-/**
- * Normalize the chart's primary source records across supported data shapes.
- * Used for count and integrity hashing only; records are never stored.
- */
-function normalizeSourceRecords(
-  component: string,
-  props: Record<string, unknown>
-): ReadonlyArray<unknown> {
-  if (Array.isArray(props.data)) return props.data
-  if (
-    component === "ForceDirectedGraph" ||
-    component === "SankeyDiagram" ||
-    component === "ProcessSankey" ||
-    component === "ChordDiagram"
-  ) {
-    return [
-      ...(Array.isArray(props.nodes) ? props.nodes : []),
-      ...(Array.isArray(props.edges) ? props.edges : [])
-    ]
-  }
-  if (component === "ChoroplethMap") {
-    if (Array.isArray(props.areas)) return props.areas
-    if (
-      props.areas &&
-      typeof props.areas === "object" &&
-      Array.isArray((props.areas as Record<string, unknown>).features)
-    ) {
-      return (props.areas as { features: unknown[] }).features
-    }
-    return typeof props.areas === "string"
-      ? [{ geographyReference: props.areas }]
-      : []
-  }
-  if (component === "ProportionalSymbolMap") {
-    return Array.isArray(props.points) ? props.points : []
-  }
-  if (component === "FlowMap") {
-    return [
-      ...(Array.isArray(props.nodes) ? props.nodes : []),
-      ...(Array.isArray(props.flows) ? props.flows : [])
-    ]
-  }
-  if (component === "DistanceCartogram") {
-    return [
-      ...(Array.isArray(props.points) ? props.points : []),
-      ...(Array.isArray(props.lines) ? props.lines : [])
-    ]
-  }
-  if (
-    component === "TreeDiagram" ||
-    component === "Treemap" ||
-    component === "CirclePack" ||
-    component === "OrbitDiagram"
-  ) {
-    return props.data && typeof props.data === "object" ? [props.data] : []
-  }
-  if (component === "BigNumber" || component === "GaugeChart") {
-    const value = props.value
-    return value === undefined ? [] : [{ value }]
-  }
-  return []
-}
-
 /** Assemble a versioned evidence envelope from existing Semiotic payloads. */
 export function toEvidenceEnvelope(
   component: string,
@@ -515,37 +420,6 @@ export function toEvidenceEnvelope(
   }
 }
 
-/** Validate the accessibility fields consumed by the publication gate. */
-function validateEnvelopeAccessibilityAudit(value: unknown, path: string): void {
-  if (value === undefined) return
-  if (!value || typeof value !== "object" || Array.isArray(value)) {
-    throw new TypeError(`${path} must be an accessibility audit object`)
-  }
-  const audit = value as Record<string, unknown>
-  if (audit.ok !== undefined && typeof audit.ok !== "boolean") {
-    throw new TypeError(`${path}.ok must be boolean`)
-  }
-  if (audit.findings === undefined) return
-  if (!Array.isArray(audit.findings)) {
-    throw new TypeError(`${path}.findings must be an array`)
-  }
-  for (const [index, finding] of audit.findings.entries()) {
-    const findingPath = `${path}.findings[${index}]`
-    if (!finding || typeof finding !== "object" || Array.isArray(finding)) {
-      throw new TypeError(`${findingPath} must be a finding object`)
-    }
-    if (finding.critical !== undefined && typeof finding.critical !== "boolean") {
-      throw new TypeError(`${findingPath}.critical must be boolean`)
-    }
-    if (
-      finding.status !== undefined &&
-      !["pass", "fail", "warn", "manual", "not-applicable"].includes(finding.status)
-    ) {
-      throw new TypeError(`${findingPath}.status must be an accessibility finding status`)
-    }
-  }
-}
-
 /** Validate and restore an envelope without silently trusting malformed input. */
 export function fromEvidenceEnvelope(value: unknown): ChartEvidenceEnvelope {
   if (!value || typeof value !== "object" || Array.isArray(value)) {
@@ -625,25 +499,12 @@ export function fromEvidenceEnvelope(value: unknown): ChartEvidenceEnvelope {
   if (!["svg", "canvas", "png", "not-rendered"].includes(render.mode)) {
     throw new TypeError("Evidence envelope render requires mode")
   }
-  for (const section of [render, render.evidence]) {
-    if (
-      section?.sceneHashVersion !== undefined &&
-      (section.sceneHashVersion !== 2 ||
-        typeof section.sceneHash !== "string" ||
-        !/^[a-f0-9]{64}$/.test(section.sceneHash))
-    ) {
-      throw new TypeError("Evidence envelope requires a supported scene hash version and SHA-256 digest")
-    }
-  }
-  if (
-    render.evidence &&
-    (render.sceneHashVersion === 2 || render.evidence.sceneHashVersion === 2) &&
-    (render.sceneHashVersion !== 2 ||
-      render.evidence.sceneHashVersion !== 2 ||
-      render.sceneHash !== render.evidence.sceneHash)
-  ) {
-    throw new TypeError("Evidence envelope scene hash does not match its render evidence")
-  }
+  validateEnvelopeSceneHash(render)
+  validateEnvelopeArtifactAttachment({
+    contract: render.evidence?.artifactContract,
+    transfer: render.evidence?.artifactTransfer,
+    binding: render.evidence?.artifactBinding
+  }, "render.evidence.artifact")
   const modality =
     envelope.modalityChecks as ChartEvidenceEnvelope["modalityChecks"]
   for (const section of ["structured", "vision", "tandem"] as const) {
@@ -694,33 +555,16 @@ export function fromEvidenceEnvelope(value: unknown): ChartEvidenceEnvelope {
       )
     }
     const artifact = envelope.artifact as EnvelopeArtifactSection
-    const transfer = artifact.transfer
-    const binding = artifact.identityBinding
-    if (
-      !["preserved", "unsupported-version", "invalid", "excluded"].includes(
-        transfer.status
-      ) ||
-      !Array.isArray(transfer.omittedPaths) ||
-      !transfer.omittedPaths.every((path) => typeof path === "string") ||
-      !Array.isArray(transfer.warnings) ||
-      !transfer.warnings.every((warning) => typeof warning === "string")
-    ) {
-      throw new TypeError(
-        "Evidence envelope artifact has an invalid transfer report"
-      )
-    }
+    validateEnvelopeArtifactAttachment({
+      contract: artifact.contract,
+      transfer: artifact.transfer,
+      binding: artifact.identityBinding
+    }, "artifact")
     const currentBinding =
       artifact.transferBindingVersion === ARTIFACT_TRANSFER_BINDING_VERSION
-    const validIdentityBinding =
-      binding &&
-      ["match", "mismatch", "unknown"].includes(binding.status) &&
-      Array.isArray(binding.mismatchPaths) &&
-      binding.mismatchPaths.every((path) => typeof path === "string") &&
-      Array.isArray(binding.unknownPaths) &&
-      binding.unknownPaths.every((path) => typeof path === "string")
     if (
-      (currentBinding && !validIdentityBinding) ||
-      (!currentBinding && binding !== undefined)
+      (currentBinding && artifact.identityBinding === undefined) ||
+      (!currentBinding && artifact.identityBinding !== undefined)
     ) {
       throw new TypeError(
         "Evidence envelope artifact has an invalid identity binding"
@@ -742,22 +586,6 @@ export function fromEvidenceEnvelope(value: unknown): ChartEvidenceEnvelope {
       throw new TypeError(
         "Evidence envelope artifact transfer fingerprint does not match its payload"
       )
-    }
-    if (envelope.artifact.contract !== undefined) {
-      const restored = serializeArtifactContract(envelope.artifact.contract)
-      const declaredStatus = transfer.status
-      const compatibleExcludedStatus =
-        declaredStatus === "excluded" &&
-        restored.transfer.status === "preserved" &&
-        transfer.omittedPaths.length > 0
-      if (
-        restored.transfer.status !== declaredStatus &&
-        !compatibleExcludedStatus
-      ) {
-        throw new TypeError(
-          "Evidence envelope artifact transfer status does not match its contract"
-        )
-      }
     }
   }
   for (const conflict of modality.tandem.conflicts) {
